Add files via upload
enrollment_processor.py uses the Selenium Webdriver for Python, instead of the python-mechanize module. So far, this script only opens up Firefox and googles for "Selenium". I had to change a few settings to get it to work with Firefox (the Python Selenium Webdriver only works with Firefox 45), but it's easy to take those lines out and use Chrome, instead.